+++ consultant/src/main/java/com/huha/consultant/ConsultantApplication.java
@@ -1,0 +1,13 @@
+package com.huha.consultant;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ConsultantApplication {
+
+    public static void main(String[] args) {
+        SpringApplication.run(ConsultantApplication.class, args);
+    }
+
+}
