+++ consultant/src/main/java/com/huha/consultant/aiservice/ConsultantService.java
@@ -1,0 +1,26 @@
+package com.huha.consultant.aiservice;
+
+import dev.langchain4j.service.MemoryId;
+import dev.langchain4j.service.SystemMessage;
+import dev.langchain4j.service.UserMessage;
+import dev.langchain4j.service.spring.AiService;
+import dev.langchain4j.service.spring.AiServiceWiringMode;
+import reactor.core.publisher.Flux;
+
+@AiService(
+        wiringMode = AiServiceWiringMode.EXPLICIT,
+        chatModel = "openAiChatModel",
+        streamingChatModel = "openAiStreamingChatModel",
+//        chatMemory = "chatMemory"//配置会话记忆对象
+        chatMemoryProvider = "chatMemoryProvider",//配置会话记忆对象提供者
+        contentRetriever = "contentRetriever",//配置向量数据库检索对象
+        tools = "reservationTool"
+)
+//@AiService
+public interface ConsultantService {
+    //用于聊天的的方法
+//    @SystemMessage("你是东哥的小月月，人美心善又多金？")
+//    @UserMessage()
+    @SystemMessage(fromResource = "system.txt")
+    public Flux<String> chat(@MemoryId String memoryId, @UserMessage String message);
+}
