+++ consultant/src/main/java/com/huha/consultant/pojo/Reservation.java
@@ -1,0 +1,20 @@
+package com.huha.consultant.pojo;
+
+import lombok.AllArgsConstructor;
+import lombok.Data;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Data
+@NoArgsConstructor
+@AllArgsConstructor
+public class Reservation {
+    private Long id;
+    private String name;
+    private String gender;
+    private String phone;
+    private LocalDateTime communicationTime;
+    private String province;
+    private Integer estimatedScore;
+}
