--- consultant/src/main/java/com/itheima/consultant/ConsultantApplication.java
@@ -1,13 +1,0 @@
-package com.itheima.consultant;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ConsultantApplication {
-
-    public static void main(String[] args) {
-        SpringApplication.run(ConsultantApplication.class, args);
-    }
-
-}
